consolidate db_utils queries into config.py, transfer db_util functions into utils.py and remove db_utils.py
